quanxian

+++ app/Model/Menu.php
@@ -1,26 +1,25 @@
 <?php
 
-
 namespace App\Model;
 
-
+use App\Http\Requests\administrator\MenuRequest;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Support\Facades\Hash;
 
-class User extends Model
+class Menu extends Model
 {
-    /** 用户 */
+    /** 菜单*/
     /**
-     *用户创建
+     *菜单创建
      * @param $data
      * @return array
      */
     public function create($data)
     {
         try{
-            $data['password'] = Hash::make($data['password']);
+
             $resu=static::insert($data);
-            return isset($resu) ?  result('SUCCESS','用户创建成功') :  result('SUCCESS','用户创建失败');
+
+            return !empty($resu) ?  result('SUCCESS','菜单创建成功') :  result('SUCCESS','菜单创建失败');
 
         }catch (\Exception $exception){
             exception($exception,'VideoResource');
@@ -29,15 +28,19 @@ class User extends Model
     }
 
     /**
-     * 删除用户
+     * 删除菜单
      * @param $id
      * @return array
      */
     public function del($id)
     {
         try {
+            $data = static::where('parent_id', $id)->first();
+            if ($data) return result('FALT', '您的菜单下还有子菜单，请先删除子菜单');
+
             $start = static::where('id', $id)->delete();
-            return isset($start) ? result('SUCCESS','用户删除成功') : result('FALT','用户删除失败,请您刷新后重试');
+           return isset($start) ? result('SUCCESS','菜单删除成功') : result('FALT','菜单删除失败,请您刷新后重试');
+
         }catch (\Exception $exception){
             exception($exception,'VideoResource');
             return result('FAIL',$exception->getMessage());
@@ -45,7 +48,7 @@ class User extends Model
     }
 
     /**
-     * 获取用户
+     * 获取菜单
      */
     public function list()
     {
